Enum mentat_query_sql::ColumnOrExpression

pub enum ColumnOrExpression {
    Column(QualifiedAlias),
    ExistingColumn(Name),
    Entid(Entid),
    Integer(i32),
    Long(i64),
    Value(TypedValue),
    NullableAggregate(Box<Expression>, ValueType),
    Expression(Box<Expression>, ValueType),
}

One of the things that can appear in a projection or a constraint. Note that we use TypedValue here; it's not pure SQL, but it avoids us having to concern ourselves at this point with the translation between a TypedValue and the storage-layer representation.

Eventually we might allow different translations by providing a different QueryBuilder implementation for each storage backend. Passing TypedValues here allows for that.

Variants
Column(QualifiedAlias)ExistingColumn(Name)Entid(Entid)Integer(i32)Long(i64)Value(TypedValue)NullableAggregate(Box<Expression>, ValueType)Expression(Box<Expression>, ValueType)